Moved GetPlatform and ConvertFrom files into their own directories
