+++ frontend/src/save-formats/Wii/GetPlatform/HttpClient.js
