+++ frontend/tests/unit/save-formats/Wii/GetPlatform/MockHttpClient.js
